extracted report modal for comment
